refactor codebase into multiple files
